rename portable stories files, refactor tests
