+++ code/renderers/vue3/src/__tests__/composeStories/Button.stories.ts
@@ -53,7 +53,6 @@ const getCaptionForLocale = (locale: string) => {
 export const CSF2StoryWithLocale: CSF2Story = (args, { globals }) => ({
   components: { Button },
   setup() {
-    console.log({ globals });
     const label = getCaptionForLocale(globals.locale);
     return { args: { ...args, label } };
   },
