+++ code/renderers/vue3/src/__tests__/composeStories/portable-stories.test.ts
@@ -1,8 +1,10 @@
 /// <reference types="@testing-library/jest-dom" />;
 import { it, expect, vi, describe } from 'vitest';
 import { render, screen } from '@testing-library/vue';
+import { addons } from '@storybook/preview-api';
 import { expectTypeOf } from 'expect-type';
 import type { Meta } from '@storybook/vue3';
+
 import * as stories from './Button.stories';
 import type Button from './Button.vue';
 import { composeStories, composeStory, setProjectAnnotations } from '../../portable-stories';
@@ -91,6 +93,21 @@ describe('CSF3', () => {
   });
 });
 
+// common in addons that need to communicate between manager and preview
+it('should pass with decorators that need addons channel', () => {
+  const PrimaryWithChannels = composeStory(stories.CSF3Primary, stories.default, {
+    decorators: [
+      (StoryFn: any) => {
+        addons.getChannel();
+        return StoryFn();
+      },
+    ],
+  });
+  render(PrimaryWithChannels({ label: 'Hello world' }));
+  const buttonElement = screen.getByText(/Hello world/i);
+  expect(buttonElement).not.toBeNull();
+});
+
 describe('ComposeStories types', () => {
   it('Should support typescript operators', () => {
     type ComposeStoriesParam = Parameters<typeof composeStories>[0];
@@ -105,4 +122,17 @@ describe('ComposeStories types', () => {
       default: stories.default satisfies Meta<typeof Button>,
     }).toMatchTypeOf<ComposeStoriesParam>();
   });
+});
+
+// Batch snapshot testing
+const testCases = Object.values(composeStories(stories)).map((Story) => [Story.storyName, Story]);
+it.each(testCases)('Renders %s story', async (_storyName, Story) => {
+  if (typeof Story === 'string' || _storyName === 'CSF2StoryWithParamsAndDecorator') {
+    return;
+  }
+
+  await new Promise((resolve) => setTimeout(resolve, 0));
+
+  const tree = await render(Story());
+  expect(tree.baseElement).toMatchSnapshot();
 });
